refactor(suitability-triage): extract supersession-detection kernel
evaluateHighConfidenceDuplicate and its three gh argv-builders
(buildClosedByMergedPrArgs, buildMergedPrListArgs, buildPrFilesArgs)
implemented a general-purpose "is this issue already superseded by a
merged PR" algorithm but lived entirely inside suitability-triage.mts,
whose public contract is Check 4's shape -- so a future B2.0
mechanization of the post-claim supersession re-check would either
duplicate this kernel or reach across script boundaries.

Move them, plus their supporting types and the CheckOutcome shape, into
a new src/scripts/supersession-detection.mts module that
suitability-triage.mts now imports from. The module carries no import
back to suitability-triage.mts, so a future consumer never needs to
pull in the whole 7-check CLI to reuse the kernel.

Also close a real correctness-drift bug: suitability-triage.mts's
loadHighContentionFiles() hardcoded DEFAULT_MANIFEST_PATH and
DEFAULT_BUNDLE_IDS with no override, so a repository that customizes
its A4 Step 2 high-contention bundles silently got a stale Check-4
exclusion set. Add --manifest/--bundles CLI flags mirroring
discover-shared-file-overlap.mjs's own, and parameterize
loadHighContentionFiles so it validates completeness against the
requested bundle IDs instead of the hardcoded default.

Finally, add a typed tier: 'high-confidence' | 'weak' field to the
check-result shape so a future consumer can branch on Check 4's
confidence level without parsing evidence prose.

#1484's detect-only outcome/evidence contract is unchanged: existing
suitability-triage.test.mts and discover-shared-file-overlap.test.mts
assertions pass unmodified (only import paths moved for the relocated
evaluateHighConfidenceDuplicate/argv-builder tests, now in a new
supersession-detection.test.mts). New tests cover the extraction, the
override surface (including a structural pin on the runCli wiring
call site), and the tier field.

Refs #1484, #1815
