+++ .gitattributes
@@ -71,6 +71,7 @@ scripts/review-disposition-verify.mjs linguist-generated=true
 scripts/select-desynced-index.mjs linguist-generated=true
 scripts/stalled-session-quiet-check.mjs linguist-generated=true
 scripts/suitability-triage.mjs linguist-generated=true
+scripts/supersession-detection.mjs linguist-generated=true
 scripts/sync-docs.mjs linguist-generated=true
 scripts/update-fixtures.mjs linguist-generated=true
 scripts/validate-schemas.mjs linguist-generated=true
