+++ src/scripts/suitability-triage.mts
@@ -11,59 +11,38 @@ import { resolve } from 'node:path';
 import { parseCliArgs } from './cli-args.mts';
 import {
   DEFAULT_BUNDLE_IDS,
-  DEFAULT_EXTRA_FILES,
   DEFAULT_MANIFEST_PATH,
   ghJson as ghJsonArray,
-  normalizeContentionPath,
   parseCandidateFiles,
   resolveHighContentionFiles,
 } from './discover-shared-file-overlap.mts';
 import { GH_TEXT_LOOP_TIMEOUT_OPTIONS, ghText } from './gh-exec.mts';
 import { loadPolicyConfig } from './idd-config.mts';
 import { normalizePolicyConfig, POLICY_DEFAULTS } from './policy-helpers.mts';
-
-/** Upper bound on the #1484 bounded merged-PR scan (mirrors B2.0's own
- * documented `gh pr list --limit 50`). */
-const MERGED_PR_SCAN_LIMIT = 50;
+import {
+  buildClosedByMergedPrArgs,
+  buildMergedPrListArgs,
+  buildPrFilesArgs,
+  type CheckOutcome,
+  evaluateHighConfidenceDuplicate,
+  type HighConfidenceDuplicateInput,
+  type HighConfidenceMergedPr,
+} from './supersession-detection.mts';
 
 /**
  * Wall-clock budget for the #1484 merged-PR file-overlap scan (CodeRabbit
- * review finding on this PR): up to MERGED_PR_SCAN_LIMIT sequential
- * `gh pr view` calls at 30s each could otherwise take ~25 minutes in the
- * worst case (a degraded/rate-limited GitHub API). Stop early and return
- * whatever has been collected once this budget elapses, rather than
- * blocking the whole A4.5 evaluation on a slow scan. Referenced from inside
- * `fetchMergedPrFileOverlapEvidence`, which the `import.meta.main` trigger
- * below can reach synchronously, so this must stay declared above that
- * trigger for the same temporal-dead-zone reason as `CLOSED_BY_MERGED_PR_QUERY`.
+ * review finding on this PR): up to `supersession-detection.mts`'s own
+ * merged-PR-scan limit (50, mirroring B2.0's own documented `gh pr list
+ * --limit 50`) sequential `gh pr view` calls at 30s each could otherwise
+ * take ~25 minutes in the worst case (a degraded/rate-limited GitHub API).
+ * Stop early and return whatever has been collected once this budget
+ * elapses, rather than blocking the whole A4.5 evaluation on a slow scan.
+ * (#1499: that limit is baked into `buildMergedPrListArgs`'s own argv,
+ * which this file now only calls rather than builds -- this comment stays
+ * prose-only rather than importing the value, since nothing here needs it
+ * as a live binding.)
  */
 const MERGED_PR_SCAN_DEADLINE_MS = 2 * 60 * 1000;
-
-/**
- * GraphQL query for the closed-by-merged-PR read (#1484), bounded to the
- * first 50 closing-PR references. A later page could theoretically hold an
- * additional MERGED reference this misses, but that only makes the tier
- * under-detect (fall back to the weak heuristic) rather than over-detect --
- * the safe fail direction for a check that must never fail TOWARD a false
- * positive, so a full pagination loop (as `idd-roadmap-audit-execute.mts`'s
- * `hasOpenClosingPr` implements for its own different, block-a-close use
- * case) is not required here. Declared here, above the `import.meta.main`
- * trigger below, rather than alongside the other #1484 CLI glue further
- * down: the trigger calls `runCli()` synchronously at module-evaluation
- * time, and a `const` declared after that point is still in the temporal
- * dead zone when the trigger fires (see `discover-shared-file-overlap.mts`'s
- * identical note on its own flag-spec constant).
- */
-const CLOSED_BY_MERGED_PR_QUERY = `query($owner:String!,$repo:String!,$number:Int!){
-  repository(owner:$owner,name:$repo){
-    issue(number:$number){
-      state
-      closedByPullRequestsReferences(first:50){
-        nodes { number state }
-      }
-    }
-  }
-}`;
 
 /** Parsed CLI arguments. */
 interface SuitabilityTriageArgs {
@@ -72,6 +51,15 @@ interface SuitabilityTriageArgs {
   owner: string;
   repo: string;
   policy: string;
+  /** #1499: high-contention manifest override, mirroring
+   * `discover-shared-file-overlap.mts`'s own `--manifest` flag so a
+   * repository that customizes its A4 Step 2 manifest path gets a matching
+   * Check-4 exclusion set instead of the hardcoded default. */
+  manifest: string;
+  /** #1499: high-contention bundle-id override, mirroring
+   * `discover-shared-file-overlap.mts`'s own `--bundles` flag. `null` means
+   * "not passed" -- the caller falls back to `DEFAULT_BUNDLE_IDS`. */
+  bundles: string[] | null;
   verbose: boolean;
   help: boolean;
 }
@@ -95,6 +83,8 @@ const SUITABILITY_TRIAGE_FLAG_SPEC = {
   '--owner': { type: 'string', default: '' },
   '--repo': { type: 'string', default: '' },
   '--policy': { type: 'string', default: '' },
+  '--manifest': { type: 'string', default: DEFAULT_MANIFEST_PATH },
+  '--bundles': { type: 'string' },
   '--verbose': { type: 'boolean', default: false },
   '--help': { type: 'boolean', short: 'h' },
 } as const;
@@ -120,36 +110,6 @@ interface DuplicateCandidate {
   title: string;
   state: string;
   url: string;
-}
-
-/** One merged PR's changed-file evidence for the high-confidence tier (#1484). */
-export interface HighConfidenceMergedPr {
-  number: number;
-  mergedAt: string;
-  files: string[];
-}
-
-/**
- * Mechanical B2.0-style evidence for the Check 4 high-confidence tier
- * (#1484): a candidate issue's own `closedByPullRequestsReferences`, plus a
- * bounded merged-PR-vs-`## Candidate files` overlap scan. Reuses
- * `discover-shared-file-overlap.mts`'s path normalization and high-contention
- * set instead of re-implementing either. Every field is defensively
- * re-validated by `evaluateHighConfidenceDuplicate` itself (not just at the
- * `evaluateSuitability` options boundary), so a caller that hand-builds a
- * `Context` directly -- as several existing tests already do, bypassing
- * normalization -- can never crash it or manufacture a false hit from a
- * malformed shape; a missing or malformed field just falls through to
- * today's weak heuristic unchanged.
- */
-export interface HighConfidenceDuplicateInput {
-  closedByMergedPrNumbers: number[];
-  candidateFiles: string[];
-  /** High-contention files (bundle + manifest) excluded from the overlap
-   * check -- a coincidental hit on a broadly-shared file is not on its own
-   * high-confidence evidence that THIS issue was superseded. */
-  highContentionFiles: string[];
-  mergedPrs: HighConfidenceMergedPr[];
 }
 
 interface Context {
@@ -179,16 +139,15 @@ interface Context {
   highConfidenceCollectionDegraded?: boolean;
 }
 
-interface CheckOutcome {
-  pass: boolean;
-  evidence: string;
-}
-
 interface CheckResult {
   id: string;
   name: string;
   result: string;
   evidence: string;
+  /** #1499: present only for a fail whose evidence came from the
+   * high-confidence mechanical kernel vs. the weak title/declaration
+   * heuristic -- see `CheckOutcome` in `supersession-detection.mts`. */
+  tier?: 'high-confidence' | 'weak';
 }
 
 interface SuitabilityResult {
@@ -367,6 +326,7 @@ export function evaluateSuitability(
       name: check.name,
       result: result.pass ? 'pass' : 'fail',
       evidence: result.evidence,
+      ...(result.tier ? { tier: result.tier } : {}),
     });
     if (!result.pass) {
       return {
@@ -561,86 +521,6 @@ export function checkTrustSafety(context: Context): CheckOutcome {
   };
 }
 
-/**
- * High-confidence Check 4 tier (#1484): evaluate the mechanical B2.0-style
- * signals -- a merged closing-PR reference on the candidate issue itself, or
- * a merged PR that already changed one of the issue's own declared
- * `## Candidate files` (excluding high-contention/shared files, which many
- * unrelated issues touch and so are not on their own high-confidence
- * evidence that THIS issue was superseded). Returns `null` -- never a
- * synthesized verdict of its own -- whenever no strong signal fires, so the
- * caller falls through to the existing weak title/declaration heuristic
- * unchanged. This is the fail-safe contract the issue requires: never fail
- * TOWARD a false high-confidence flag. `input` may be `undefined` (evidence
- * not collected by the caller) or a partially malformed shape; both degrade
- * to "no verdict" rather than a crash or a false hit.
- */
-export function evaluateHighConfidenceDuplicate(
-  input: HighConfidenceDuplicateInput | undefined,
-): CheckOutcome | null {
-  if (!input) {
-    return null;
-  }
-
-  const closedByMergedPrNumbers = (
-    Array.isArray(input.closedByMergedPrNumbers)
-      ? input.closedByMergedPrNumbers
-      : []
-  ).filter((n) => Number.isInteger(n) && n > 0);
-  if (closedByMergedPrNumbers.length > 0) {
-    return {
-      pass: false,
-      evidence: `High-confidence duplicate: issue is already referenced by merged closing PR(s) #${closedByMergedPrNumbers.join(', #')} (closedByPullRequestsReferences).`,
-    };
-  }
-
-  const highContention = new Set(
-    (Array.isArray(input.highContentionFiles)
-      ? input.highContentionFiles
-      : []
-    ).map((file) => normalizeContentionPath(file)),
-  );
-  const candidateFiles = [
-    ...new Set(
-      (Array.isArray(input.candidateFiles) ? input.candidateFiles : [])
-        .map((file) => normalizeContentionPath(file))
-        .filter((file) => file.length > 0 && !highContention.has(file)),
-    ),
-  ];
-  if (candidateFiles.length === 0) {
-    return null;
-  }
-  const candidateSet = new Set(candidateFiles);
-
-  const mergedPrs: unknown[] = Array.isArray(input.mergedPrs)
-    ? input.mergedPrs
-    : [];
-  for (const raw of mergedPrs) {
-    const pr = (raw ?? {}) as {
-      number?: unknown;
-      mergedAt?: unknown;
-      files?: unknown;
-    };
-    const number = Number(pr.number);
-    if (!Number.isInteger(number) || number <= 0) {
-      continue;
-    }
-    const files = Array.isArray(pr.files) ? pr.files : [];
-    const overlap = [
-      ...new Set(files.map((file) => normalizeContentionPath(file))),
-    ].filter((file) => candidateSet.has(file));
-    if (overlap.length > 0) {
-      const mergedAt = String(pr.mergedAt ?? '');
-      return {
-        pass: false,
-        evidence: `High-confidence duplicate: merged PR #${number}${mergedAt ? ` (merged ${mergedAt})` : ''} already changed candidate file(s): ${overlap.sort().join(', ')}.`,
-      };
-    }
-  }
-
-  return null;
-}
-
 export function checkDuplicateOrSuperseded(context: Context): CheckOutcome {
   const highConfidence = evaluateHighConfidenceDuplicate(
     context.highConfidenceDuplicate,
@@ -670,6 +550,7 @@ export function checkDuplicateOrSuperseded(context: Context): CheckOutcome {
       return {
         pass: false,
         evidence: `Exact-title duplicate found: #${degradedExactMatch.number}`,
+        tier: 'weak',
       };
     }
     return {
@@ -692,6 +573,7 @@ export function checkDuplicateOrSuperseded(context: Context): CheckOutcome {
     return {
       pass: false,
       evidence: `Issue body declares duplicate/superseded status: ${matched}`,
+      tier: 'weak',
     };
   }
 
@@ -706,6 +588,7 @@ export function checkDuplicateOrSuperseded(context: Context): CheckOutcome {
     return {
       pass: false,
       evidence: `Exact-title duplicate found: #${duplicate.number}`,
+      tier: 'weak',
     };
   }
 
@@ -724,6 +607,7 @@ export function checkDuplicateOrSuperseded(context: Context): CheckOutcome {
     return {
       pass: false,
       evidence: `Near-duplicate found: #${nearDuplicate.number} ("${nearDuplicate.title}"). Title similarity >80%.`,
+      tier: 'weak',
     };
   }
 
@@ -1014,7 +898,12 @@ function runCli(): void {
     // against -- most issues have none, and the result would otherwise be
     // discarded.
     const resolvedHighContentionFiles =
-      candidateFiles.length > 0 ? loadHighContentionFiles() : null;
+      candidateFiles.length > 0
+        ? loadHighContentionFiles(
+            args.manifest,
+            args.bundles ?? DEFAULT_BUNDLE_IDS,
+          )
+        : null;
     const shouldScanMergedPrs =
       candidateFiles.length > 0 &&
       resolvedHighContentionFiles !== null &&
@@ -1094,6 +983,10 @@ function runCli(): void {
           id: check.id,
           name: check.name,
           result: check.result,
+          // #1499: carried through even in non-verbose mode -- the typed
+          // tier signal exists precisely so a consumer can branch on it
+          // without asking for full evidence prose.
+          ...(check.tier ? { tier: check.tier } : {}),
         })),
   };
 
@@ -1130,6 +1023,17 @@ export function parseArgs(argv: string[]): SuitabilityTriageArgs {
     owner: values.owner as string,
     repo: values.repo as string,
     policy: values.policy as string,
+    manifest: values.manifest as string,
+    // #1499: mirrors `discover-shared-file-overlap.mts`'s own `--bundles`
+    // parsing exactly -- absent means "not passed" (`null`), present is a
+    // comma-split, trimmed, empty-token-filtered list.
+    bundles:
+      values.bundles === undefined
+        ? null
+        : String(values.bundles as string)
+            .split(',')
+            .map((part) => part.trim())
+            .filter(Boolean),
     verbose: values.verbose as boolean,
     help,
   };
@@ -1149,7 +1053,13 @@ function loadPolicy(policyPath: string): unknown {
 
 function printHelp(): void {
   process.stdout.write(`Usage:
-  node scripts/suitability-triage.mjs --issue <number> [--token <token>] [--owner <owner>] [--repo <repo>] [--policy <path>] [--verbose] [--help]
+  node scripts/suitability-triage.mjs --issue <number> [--token <token>] [--owner <owner>] [--repo <repo>] [--policy <path>] [--manifest <path>] [--bundles <id1,id2>] [--verbose] [--help]
+
+--manifest / --bundles override the Check 4 high-confidence tier's
+high-contention exclusion set (default: the same manifest path and bundle
+IDs as discover-shared-file-overlap.mjs's own --manifest/--bundles), so a
+repository that customizes its A4 Step 2 contention bundles gets a matching
+Check-4 exclusion set instead of a stale hardcoded default.
 
 Output schema:
 {
@@ -1158,7 +1068,7 @@ Output schema:
   "passed": true,
   "outcome": "ready|unclear|needs-decision|blocked-by-human|duplicate|out-of-scope|invalid",
   "failedCheck": "repository_fit|...|null",
-  "checks": [{"id":"repository_fit","name":"Repository Fit","result":"pass|fail","evidence":"..."}]
+  "checks": [{"id":"repository_fit","name":"Repository Fit","result":"pass|fail","evidence":"...","tier":"high-confidence|weak"}]
 }
 `);
 }
@@ -1352,84 +1262,12 @@ function fetchDuplicateCandidates(
 }
 
 // --- #1484: high-confidence Check 4 tier CLI glue ---------------------------
-// Read-only: every function below only ever calls `gh api graphql` (with a
-// `query` operation, never `mutation`), `gh pr list`, or `gh pr view` (no
-// -X/--method, no issue/PR mutation subcommand).
-// #1484 is detect-only by design; do not add a mutating gh call here -- a
-// later gated-close follow-up (#1485) is a separate, human-gated change.
-// The argv-builders are exported so tests can assert the exact read-only
-// verb without shelling out (a compiled-text grep for mutating verb
-// literals would miss a `gh api ... -X POST`-shaped mutation, since none of
-// this file's own calls use one).
-
-/**
- * Argv for the closed-by-merged-PR read. Uses `gh api graphql` rather than
- * `gh issue view --json closedByPullRequestsReferences`: the latter's
- * REST-shimmed shape carries no per-PR `state`, and the connection includes
- * OPEN (not yet merged) PRs, not only merged ones -- confirmed empirically
- * against this repo's own issue #1489 (OPEN) / PR #1497 (OPEN), and matches
- * `idd-roadmap-audit-execute.mts`'s documented note that the field "returns
- * merged PRs even with `includeClosedPrs:false`" (i.e. state alone
- * determines relevance, not that flag). Filtering to `state === 'MERGED'`
- * happens in `fetchClosedByMergedPrNumbers` below, after this fetch --
- * without it, an issue with only an in-progress unmerged closing PR would
- * wrongly read as "already referenced by a merged closing PR".
- */
-export function buildClosedByMergedPrArgs(
-  owner: string,
-  repo: string,
-  issueNumber: number,
-): string[] {
-  return [
-    'api',
-    'graphql',
-    '-f',
-    `query=${CLOSED_BY_MERGED_PR_QUERY}`,
-    '-f',
-    `owner=${owner}`,
-    '-f',
-    `repo=${repo}`,
-    '-F',
-    `number=${issueNumber}`,
-  ];
-}
-
-/** Argv for the bounded merged-PR list scan (mirrors B2.0's own documented
- * `gh pr list --search "merged:>=<since>"` shape). */
-export function buildMergedPrListArgs(
-  repoRef: string,
-  sinceIso: string,
-): string[] {
-  return [
-    'pr',
-    'list',
-    '--repo',
-    repoRef,
-    '--state',
-    'merged',
-    '--search',
-    `merged:>=${sinceIso}`,
-    '--json',
-    'number,mergedAt',
-    '--limit',
-    String(MERGED_PR_SCAN_LIMIT),
-  ];
-}
-
-/** Argv for one merged PR's changed-file list. */
-export function buildPrFilesArgs(repoRef: string, prNumber: number): string[] {
-  return [
-    'pr',
-    'view',
-    String(prNumber),
-    '--repo',
-    repoRef,
-    '--json',
-    'files',
-    '--jq',
-    '.files[].path',
-  ];
-}
+// The pure argv-builders (`buildClosedByMergedPrArgs`, `buildMergedPrListArgs`,
+// `buildPrFilesArgs`) and the evaluation kernel (`evaluateHighConfidenceDuplicate`)
+// moved to `supersession-detection.mts` (#1499); this file keeps only the
+// `gh`-executing orchestration below (fetch, try/catch, deadline budget,
+// `collectionWarnings`), which the issue does not name as part of the
+// extraction.
 
 /**
  * Fetch the candidate issue's own merged closing-PR references. Throws (via
@@ -1574,20 +1412,26 @@ function fetchMergedPrFileOverlapEvidence(
  * same way. `closedByPullRequestsReferences` is a separate, independent
  * signal and is unaffected by either fallback.
  */
-function loadHighContentionFiles(): string[] | null {
+export function loadHighContentionFiles(
+  manifestPath: string,
+  bundleIds: string[],
+): string[] | null {
   try {
     const manifest = JSON.parse(
-      readFileSync(resolve(process.cwd(), DEFAULT_MANIFEST_PATH), 'utf8'),
+      readFileSync(resolve(process.cwd(), manifestPath), 'utf8'),
     );
     // Codex P2 review finding: a manifest that parses but lacks usable
     // `bundleBudgets` entries for one or both target bundle IDs (an empty
     // object, or an older schema) doesn't throw here -- `resolveHighContentionFiles`
-    // degrades gracefully to just `DEFAULT_EXTRA_FILES` for A4 Step 2's own,
-    // lower-stakes de-prioritization use. But for this tier, an incomplete
-    // exclusion set can miss a genuinely high-contention file, so a shared
-    // bundle/instruction file could be misread as specific overlap evidence
-    // -- exactly the false high-confidence flag Check 4 must never produce.
-    // Require both configured bundle IDs to actually resolve before
+    // degrades gracefully to just the manifest path itself for A4 Step 2's
+    // own, lower-stakes de-prioritization use. But for this tier, an
+    // incomplete exclusion set can miss a genuinely high-contention file, so
+    // a shared bundle/instruction file could be misread as specific overlap
+    // evidence -- exactly the false high-confidence flag Check 4 must never
+    // produce. Require every requested bundle ID (#1499: the caller's own
+    // `--bundles` override when given, not the hardcoded default -- a
+    // repository that customizes its bundle set must have THOSE bundles
+    // validated, not `DEFAULT_BUNDLE_IDS`) to actually resolve before
     // accepting the set; otherwise treat it the same as an unreadable
     // manifest (return null, which the caller already records as a
     // collection warning and degrades to exact-title-only).
@@ -1609,7 +1453,7 @@ function loadHighContentionFiles(): string[] | null {
     // above yet still let `resolveHighContentionFiles` silently omit that
     // bundle's real shared files -- the same false-flag risk as a missing
     // bundle id entirely, so it must degrade the same way.
-    const allBundleIdsResolved = DEFAULT_BUNDLE_IDS.every((id) =>
+    const allBundleIdsResolved = bundleIds.every((id) =>
       nonEmptyFilesBundleIds.has(id),
     );
     if (!allBundleIdsResolved) {
@@ -1618,8 +1462,12 @@ function loadHighContentionFiles(): string[] | null {
     return [
       ...resolveHighContentionFiles({
         manifest,
-        bundleIds: DEFAULT_BUNDLE_IDS,
-        extraFiles: DEFAULT_EXTRA_FILES,
+        bundleIds,
+        // #1499: mirrors `discover-shared-file-overlap.mts`'s own `runCli`
+        // pattern -- the manifest path actually in use is the file reported
+        // (and matched) as high-contention, not a hardcoded default that
+        // silently stops tracking a customized manifest.
+        extraFiles: [manifestPath],
       }),
     ];
   } catch {
